+++ tests/tasking/test_decisions.py
@@ -20,7 +20,7 @@ def mockedDecisionClass() -> Decision:
     """Return reference to a minimal :class:`.Decision` class."""
 
     class MockedDecision(Decision):
-        def calculate(self, reward_matrix, **kwargs):
+        def _calculate(self, reward_matrix, visibility_matrix):
             return 3
 
     return MockedDecision
@@ -52,8 +52,11 @@ class TestDecisionBase:
     def testDecisionCall(self, mocked_decision_class: Decision):
         """Test the call function of the Decision base class."""
         reward_matrix = asarray([[10, 1.1, 0], [4, 15.9, 1], [2.1, 3, 11]])
+        visibility_matrix = asarray(
+            [[True, False, False], [False, True, False], [False, False, True]]
+        )
         decision = mocked_decision_class()
-        decision.calculate(reward_matrix)
+        decision.calculate(reward_matrix, visibility_matrix)
 
 
 class TestMyopicNaiveGreedyDecision:
@@ -69,10 +72,13 @@ class TestMyopicNaiveGreedyDecision:
         myopic_decision = MyopicNaiveGreedyDecision()
         # Create example reward matrix & expected decision matrix
         reward_matrix = asarray([[10, 1.1, 0], [4, 15.9, 11], [2.1, 3, 1]])
-        expected = [[True, False, False], [False, True, True], [False, False, False]]
+        visibility_matrix = asarray(
+            [[True, False, False], [False, True, True], [False, False, False]]
+        )
 
         # Perform decision making, test expected
-        decision = myopic_decision.calculate(reward_matrix)
+        decision = myopic_decision.calculate(reward_matrix, visibility_matrix)
+        expected = [[True, False, False], [False, True, True], [False, False, False]]
         assert array_equal(decision, expected)
         # Assert constraint that single sensor cannot be tasked more than once
         tasked_sensors = nonzero(decision)[1]
@@ -83,10 +89,12 @@ class TestMyopicNaiveGreedyDecision:
         myopic_decision = MyopicNaiveGreedyDecision()
         # Create example reward matrix & expected decision matrix
         reward_matrix = -1 * asarray([[10, 1.1, 0], [4, 15.9, 11], [2.1, 3, 1]])
-        expected = [[False, False, False], [False, False, False], [False, False, False]]
+        visibility_matrix = full((3, 3), False)
+        expected = full((3, 3), False)
 
         # Perform decision making, test expected
-        decision = myopic_decision.calculate(reward_matrix)
+        decision = myopic_decision.calculate(reward_matrix, visibility_matrix)
+        decision = decision & visibility_matrix
         assert array_equal(decision, expected)
         # Assert constraint that single sensor cannot be tasked more than once
         tasked_sensors = nonzero(decision)[1]
@@ -98,9 +106,11 @@ class TestMyopicNaiveGreedyDecision:
         # Create example reward matrix & expected decision matrix
         reward_matrix = full((3, 3), 10)
         expected = [[True, True, True], [False, False, False], [False, False, False]]
-
+        visibility_matrix = asarray(
+            [[True, True, True], [False, False, False], [False, False, False]]
+        )
         # Perform decision making, test expected
-        decision = myopic_decision.calculate(reward_matrix)
+        decision = myopic_decision.calculate(reward_matrix, visibility_matrix)
         assert array_equal(decision, expected)
         # Assert constraint that single sensor cannot be tasked more than once
         tasked_sensors = nonzero(decision)[1]
@@ -111,9 +121,10 @@ class TestMyopicNaiveGreedyDecision:
         myopic_decision = MyopicNaiveGreedyDecision()
         # Create example null reward matrix
         reward_matrix = asarray([[0, 0, 0], [0, 0, 0], [0, 0, 0]])
-
+        visibility_matrix = full((3, 3), False)
         # Perform decision making, test expected null-tasking
-        assert not np_any(myopic_decision.calculate(reward_matrix))
+        decision = myopic_decision.calculate(reward_matrix, visibility_matrix) & visibility_matrix
+        assert not np_any(decision)
 
 
 class TestMunkresDecision:
@@ -129,10 +140,13 @@ class TestMunkresDecision:
         munkres_decision = MunkresDecision()
         # Create example reward matrix & expected decision matrix
         reward_matrix = asarray([[1.1, 10, 12], [4, 15.9, 1], [2.1, 3, 11]])
+        visibility_matrix = asarray(
+            [[False, False, True], [False, True, False], [True, False, False]]
+        )
         expected = [[False, False, True], [False, True, False], [True, False, False]]
 
         # Perform decision making, test expected
-        decision = munkres_decision.calculate(reward_matrix)
+        decision = munkres_decision.calculate(reward_matrix, visibility_matrix)
         assert array_equal(decision, expected)
         # Assert constraint that single sensor cannot be tasked more than once
         tasked_sensors = nonzero(decision)[1]
@@ -143,10 +157,11 @@ class TestMunkresDecision:
         munkres_decision = MunkresDecision()
         # Create example reward matrix & expected decision matrix
         reward_matrix = -1 * asarray([[1.1, 10, 12], [4, 15.9, 1], [2.1, 3, 11]])
-        expected = [[False, False, False], [False, False, False], [False, False, False]]
+        visibility_matrix = full((3, 3), False)
+        expected = full((3, 3), False)
 
         # Perform decision making, test expected
-        decision = munkres_decision.calculate(reward_matrix)
+        decision = munkres_decision.calculate(reward_matrix, visibility_matrix) & visibility_matrix
         assert array_equal(decision, expected)
         # Assert constraint that single sensor cannot be tasked more than once
         tasked_sensors = nonzero(decision)[1]
@@ -158,22 +173,16 @@ class TestMunkresDecision:
         # Create example reward matrix & expected decision matrix
         reward_matrix = full((3, 3), 10)
         expected = [[True, False, False], [False, True, False], [False, False, True]]
+        visibility_matrix = asarray(
+            [[True, False, False], [False, True, False], [False, False, True]]
+        )
 
         # Perform decision making, test expected
-        decision = munkres_decision.calculate(reward_matrix)
+        decision = munkres_decision.calculate(reward_matrix, visibility_matrix) & visibility_matrix
         assert array_equal(decision, expected)
         # Assert constraint that single sensor cannot be tasked more than once
         tasked_sensors = nonzero(decision)[1]
         assert array_equal(unique(tasked_sensors), sort(tasked_sensors))
-
-    def testNullDecision(self):
-        """Test the call function of the MunkresDecision class."""
-        munkres_decision = MunkresDecision()
-        # Create example null reward matrix
-        reward_matrix = asarray([[0, 0, 0], [0, 0, 0], [0, 0, 0]])
-
-        # Perform decision making, test expected null-tasking
-        assert not np_any(munkres_decision.calculate(reward_matrix))
 
 
 class TestRandomDecision:
@@ -189,10 +198,11 @@ class TestRandomDecision:
         random_decision = RandomDecision(20000)
         # Create example reward matrix & expected decision matrix
         reward_matrix = asarray([[1.1, 10, 12], [4, 15.9, 1], [2.1, 3, 11]])
+        visibility_matrix = full((3, 3), True)
         expected = [[False, False, False], [False, True, False], [True, False, True]]
 
         # Perform decision making, test expected
-        decision = random_decision.calculate(reward_matrix)
+        decision = random_decision.calculate(reward_matrix, visibility_matrix)
         assert array_equal(decision, expected)
         # Assert constraint that single sensor cannot be tasked more than once
         tasked_sensors = nonzero(decision)[1]
@@ -203,10 +213,13 @@ class TestRandomDecision:
         random_decision = RandomDecision(129281)
         # Create example reward matrix & expected decision matrix
         reward_matrix = asarray([[0, 10, 12], [4, 15.9, 0], [2.1, 3, 11]])
+        visibility_matrix = asarray(
+            [[False, False, True], [True, False, False], [False, True, False]]
+        )
         expected = [[False, False, True], [True, False, False], [False, True, False]]
 
         # Perform decision making, test expected
-        decision = random_decision.calculate(reward_matrix)
+        decision = random_decision.calculate(reward_matrix, visibility_matrix)
         assert array_equal(decision, expected)
         # Assert constraint that single sensor cannot be tasked more than once
         tasked_sensors = nonzero(decision)[1]
@@ -216,10 +229,11 @@ class TestRandomDecision:
         """Test the call function of the RandomDecision class."""
         random_decision = RandomDecision(1202102120)
         # Create example null reward matrix
+        visibility_matrix = full((3, 3), False)
         reward_matrix = asarray([[0, 0, 0], [0, 0, 0], [0, 0, 0]])
 
         # Perform decision making, test expected null-tasking
-        assert not np_any(random_decision.calculate(reward_matrix))
+        assert not np_any(random_decision.calculate(reward_matrix, visibility_matrix))
 
 
 class TestAllVisibleDecision:
@@ -234,40 +248,10 @@ class TestAllVisibleDecision:
         """Test the call function of the AllVisibleDecision class."""
         all_visible_decision = AllVisibleDecision()
         # Create example reward matrix & expected decision matrix
-        reward_matrix = asarray([[1.1, 0.0, 12], [0.0, 15.9, 1], [2.1, 3, 11]])
+        reward_matrix = full((3, 3), False)
+        visibility_matrix = asarray([[True, False, True], [False, True, True], [True, True, True]])
         expected = [[True, False, True], [False, True, True], [True, True, True]]
 
         # Perform decision making, test expected
-        decision = all_visible_decision.calculate(reward_matrix)
+        decision = all_visible_decision.calculate(reward_matrix, visibility_matrix)
         assert array_equal(decision, expected)
-
-    def testNegativeDecision(self):
-        """Test the call function of the AllVisibleDecision class."""
-        all_visible_decision = AllVisibleDecision()
-        # Create example reward matrix & expected decision matrix
-        reward_matrix = -1 * asarray([[1.1, 10, 12], [4, 15.9, 1], [2.1, 3, 11]])
-        expected = [[False, False, False], [False, False, False], [False, False, False]]
-
-        # Perform decision making, test expected
-        decision = all_visible_decision.calculate(reward_matrix)
-        assert array_equal(decision, expected)
-
-    def testEqualDecision(self):
-        """Test the call function of the AllVisibleDecision class."""
-        all_visible_decision = AllVisibleDecision()
-        # Create example reward matrix & expected decision matrix
-        reward_matrix = full((3, 3), 10)
-        expected = [[True, True, True], [True, True, True], [True, True, True]]
-
-        # Perform decision making, test expected
-        decision = all_visible_decision.calculate(reward_matrix)
-        assert array_equal(decision, expected)
-
-    def testNullDecision(self):
-        """Test the call function of the AllVisibleDecision class."""
-        all_visible_decision = AllVisibleDecision()
-        # Create example null reward matrix
-        reward_matrix = asarray([[0, 0, 0], [0, 0, 0], [0, 0, 0]])
-
-        # Perform decision making, test expected null-tasking
-        assert not np_any(all_visible_decision.calculate(reward_matrix))